requirements/base.txt
Django==5.1.7
django-environ==0.13.0
psycopg2-binary==2.9.11
Pillow==12.1.0
requests==2.32.5
beautifulsoup4==4.14.2
lxml==6.0.4
